+++ src/main/java/adaptiveauth/entity/RiskLog.java
@@ -3,7 +3,7 @@ package adaptiveauth.entity;
 public class RiskLog {
 
     public String timestamp;
-    public int score;
+    public double score;
     public String level;
     public String decision;
 }
